refactor: move build course query to subdirectory
